Updated readme.md
Added information about how-to use the new feature for counting characters in teaxtarea. Moved images from rawgit (rawgit is shutting down theire services) to user-images.githubusercontent ("upload"-hack by copy-pasting image into "Issues" without posting the issue). Hopefully this "funtionality" will exist for a long time. Some of the figures/images was redesigned to clearify the relation with elements in Digiforms. And some of the text was edited to make more sense.